feat: Add Dockerfile for web app and Docker build command
This commit adds a Dockerfile for the web app and a Docker build command in the justfile. The Dockerfile sets up the environment for running the web app using the node:22-slim base image. The Docker build command builds a Docker image for the web app using the Dockerfile. These changes enable easy deployment and containerization of the web app.