feat: Refactor main.ts to use electron-store for storing last release version
The main.ts file has been refactored to use the electron-store package for storing the last release version. The previous implementation used fs module to read and write the version to a file, but this change improves the code by using a more convenient and reliable method provided by electron-store. The store instance is created and used to get and set the last release version. Additionally, console logs have been updated to provide more informative messages about the release status and download progress.

This commit improves the code structure and enhances the reliability of storing and retrieving the last release version.