refactor: Add Separator component and update npm dependencies
This commit adds a new Separator component to the web application, located at apps/web/src/components/ui/separator.tsx. The Separator component is a thin line that can be used to visually separate content. It is implemented using the @radix-ui/react-separator library.

Additionally, the commit updates the npm dependencies in the package.json file, including adding new dependencies such as "@hookform/resolvers" and "@radix-ui/react-separator". These updates ensure that the project is using the latest stable versions of the dependencies.

These changes improve the UI of the web application by providing a reusable Separator component and keeping the dependencies up to date.