feat: Add maxSelectable prop to EditionMultiSelectForm
This commit adds a new `maxSelectable` prop to the `EditionMultiSelectForm` component in order to limit the number of selectable items. If the user tries to select more items than the specified limit, a toast notification will be displayed to inform them that the maximum limit has been reached. This enhancement improves the user experience by providing better control over the selection process.