feat: Add maxSelectable prop to MultiSelectFormField
This commit adds a new `maxSelectable` prop to the `MultiSelectFormField` component, allowing the developer to limit the number of selectable items. When the user tries to select more items than the specified limit, the excess items will be ignored. This enhancement improves the user experience by providing better control over the selection process.