feat: Update AyahService to fetch from multiple URLs
This commit updates the AyahService in the Quran API to fetch data from multiple URLs. Previously, only one URL was used, but now a list of URLs is provided. If the data cannot be fetched from one URL, the service will try the next URL in the list until successful or all URLs have been attempted. This enhancement improves the reliability of data retrieval and ensures a fallback option in case of failures.