chore: Update npm dependencies
This commit updates the npm dependencies in the package.json file. It includes updates to the "@hono/node-server" package from version 1.11.2 to 1.11.3 and the "@hono/zod-openapi" package from version 0.14.2 to 0.14.3. Additionally, the "lucide-react" package is updated from version 0.394.0 to 0.395.0. These updates ensure that the project is using the latest stable versions of the dependencies, which may include bug fixes, performance improvements, and new features.